+++ src/loaders/RasterJsonLoader.js
@@ -75,8 +75,8 @@ async function initLoader(imageData) {
 }
 
 export default class RasterLoader extends JsonLoader {
-  constructor(params) {
-    super(params);
+  constructor(dataSource, params) {
+    super(dataSource, params);
     const { url, options } = params;
     if (!url && options) {
       this.url = URL.createObjectURL(new Blob([JSON.stringify(options)]));