+++ src/loaders/data-sources/index.js
@@ -1,2 +1,3 @@
 export { default as AnnDataSource } from './AnnDataSource';
 export { default as ZarrDataSource } from './ZarrDataSource';
+export { default as JsonSource } from './JsonSource';