+++ src/loaders/types.js
@@ -8,20 +8,20 @@ import CellSetsJsonLoader from './CellSetsJsonLoader';
 import AnnDataLoaders from './anndata-loaders';
 import GenomicProfilesZarrLoader from './GenomicProfilesZarrLoader';
 
-import { AnnDataSource, ZarrDataSource } from './data-sources';
+import { AnnDataSource, ZarrDataSource, JsonSource } from './data-sources';
 
 const ANNDATA = 'anndata';
 
 export const fileTypeToLoader = {
   'expression-matrix.zarr': [ZarrDataSource, MatrixZarrLoader],
-  'clusters.json': ClustersJsonAsMatrixZarrLoader,
-  'genes.json': GenesJsonAsMatrixZarrLoader,
-  'cells.json': JsonLoader,
-  'molecules.json': JsonLoader,
-  'neighborhoods.json': JsonLoader,
-  'raster.json': RasterJsonLoader,
+  'clusters.json': [JsonSource, ClustersJsonAsMatrixZarrLoader],
+  'genes.json': [JsonSource, GenesJsonAsMatrixZarrLoader],
+  'cells.json': [JsonSource, JsonLoader],
+  'molecules.json': [JsonSource, JsonLoader],
+  'neighborhoods.json': [JsonSource, JsonLoader],
+  'raster.json': [JsonSource, RasterJsonLoader],
   'raster.ome-zarr': [ZarrDataSource, OmeZarrLoader],
-  'cell-sets.json': CellSetsJsonLoader,
+  'cell-sets.json': [JsonSource, CellSetsJsonLoader],
   [`${ANNDATA}-cell-sets.zarr`]: [AnnDataSource, AnnDataLoaders.CellSetsZarrLoader],
   [`${ANNDATA}-cells.zarr`]: [AnnDataSource, AnnDataLoaders.CellsZarrLoader],
   [`${ANNDATA}-expression-matrix.zarr`]: [AnnDataSource, AnnDataLoaders.MatrixZarrLoader],